+++ lib/options/themes.dart
@@ -1,7 +1,3 @@
-// Copyright 2018 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
 import 'package:flutter/material.dart';
 
 class AnagrammaticTheme {
@@ -23,13 +19,20 @@ final AnagrammaticTheme lightTheme = AnagrammaticTheme._(
   _buildLightTheme(),
 );
 
-final String fontFamily = 'GoogleSans';
-final Color optionsAccentColor = const Color(0xFF39CEFD);
+final Color _errorColor = const Color(0xFFB00020);
+
+final String _fontFamily = 'Roboto';
 
 TextTheme _buildTextTheme(TextTheme base) {
   return base.copyWith(
     title: base.title.copyWith(
-      fontFamily: fontFamily,
+      fontFamily: _fontFamily,
+    ),
+    headline: base.headline.copyWith(
+      fontFamily: _fontFamily,
+    ),
+    subhead: base.subhead.copyWith(
+      fontFamily: _fontFamily,
     ),
   );
 }
@@ -39,10 +42,6 @@ ThemeData _buildDarkTheme() {
   Color secondaryColor = Colors.grey[600];
   Color backgroundColor = const Color(0xFF202124);
   final ThemeData base = ThemeData.dark();
-  final ColorScheme colorScheme = const ColorScheme.dark().copyWith(
-    primary: primaryColor,
-    secondary: secondaryColor,
-  );
   return base.copyWith(
     primaryColor: primaryColor,
     buttonColor: primaryColor,
@@ -51,11 +50,7 @@ ThemeData _buildDarkTheme() {
     canvasColor: backgroundColor,
     scaffoldBackgroundColor: backgroundColor,
     backgroundColor: backgroundColor,
-    errorColor: const Color(0xFFB00020),
-    buttonTheme: ButtonThemeData(
-      colorScheme: colorScheme,
-      textTheme: ButtonTextTheme.primary,
-    ),
+    errorColor: _errorColor,
     textTheme: _buildTextTheme(base.textTheme),
     primaryTextTheme: _buildTextTheme(base.primaryTextTheme),
     accentTextTheme: _buildTextTheme(base.accentTextTheme),
@@ -65,13 +60,8 @@ ThemeData _buildDarkTheme() {
 ThemeData _buildLightTheme() {
   const Color primaryColor = Color(0xFF0175c2);
   const Color secondaryColor = Color(0xFF13B9FD);
-  final ColorScheme colorScheme = const ColorScheme.light().copyWith(
-    primary: primaryColor,
-    secondary: secondaryColor,
-  );
   final ThemeData base = ThemeData.light();
   return base.copyWith(
-    colorScheme: colorScheme,
     primaryColor: primaryColor,
     buttonColor: primaryColor,
     indicatorColor: Colors.white,
@@ -81,11 +71,7 @@ ThemeData _buildLightTheme() {
     canvasColor: Colors.white,
     scaffoldBackgroundColor: Colors.white,
     backgroundColor: Colors.white,
-    errorColor: const Color(0xFFB00020),
-    buttonTheme: ButtonThemeData(
-      colorScheme: colorScheme,
-      textTheme: ButtonTextTheme.primary,
-    ),
+    errorColor: _errorColor,
     textTheme: _buildTextTheme(base.textTheme),
     primaryTextTheme: _buildTextTheme(base.primaryTextTheme),
     accentTextTheme: _buildTextTheme(base.accentTextTheme),